alert/requirements.txt
boto3==1.17.13
Flask==1.1.2
playsound==1.2.2
vext==0.7.4
pgi==0.0.11.2
pyserial
pycairo
PyGObject
